refactor: change structure
