+++ src/App.tsx
@@ -1,5 +1,5 @@
-import Search from './Components/Search/Search';
 import ErrorBoundary from './Components/Error/ErrorBoundary';
+import Main from './Pages/Main/MainPage';
 
 export type SearchProps = {
   onSearch: (searchTerm: string) => void;
@@ -25,7 +25,7 @@ const App = () => {
   return (
     <ErrorBoundary>
       <div className="wrapper_content">
-        <Search />
+        <Main />
       </div>
     </ErrorBoundary>
   );
