+++ src/index.css
@@ -102,8 +102,10 @@ input {
 }
 
 .wrapper_content {
-  width: 100%;
   display: flex;
+  flex-direction: column;
+  max-width: 100%;
+  width: 100%;
 }
 
 .card-title {
@@ -154,9 +156,9 @@ input {
   line-height: normal;
 }
 
-.content {
+.search-results {
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   max-width: 100%;
   width: 100%;
 }
